moving the active workflow execution guard to the base sample service
