+++ app/services/samples/transfer_service.rb
@@ -16,8 +16,6 @@ module Samples
   #   service = Samples::TransferService.new(namespace, user)
   #   transferred_ids = service.execute(new_project_id, sample_ids, broadcast_target)
   class TransferService < BaseSampleService # rubocop:disable Metrics/ClassLength
-    include Samples::ActiveWorkflowExecutionGuard
-
     class TransferError < StandardError
     end
